react-router and response buttons added
react-router-dom added
response buttons (like, dislike, comment) icons added
right side bar style updated
logo div added
